Extract command logic
